+++ packages/camera/camera_avfoundation/darwin/Tests/Mocks/MockCaptureDevice.swift
@@ -30,6 +30,7 @@ class MockCaptureDevice: NSObject, CaptureDevice {
   var uniqueID = ""
   var position = AVCaptureDevice.Position.unspecified
   var deviceType = AVCaptureDevice.DeviceType.builtInWideAngleCamera
+  var isVirtualDevice = false
 
   var flutterActiveFormat: CaptureDeviceFormat {
     get {